refactor: review -- improve writing to opossum files
