+++ tests/test_cli.py
@@ -11,7 +11,7 @@ from _pytest.logging import LogCaptureFixture
 from click.testing import CliRunner, Result
 
 from opossum_lib.cli import generate
-from opossum_lib.core.services.opossum_file_writer import write_opossum_file
+from opossum_lib.core.services.write_opossum_file import write_opossum_file
 from opossum_lib.shared.constants import (
     INPUT_JSON_NAME,
     OUTPUT_JSON_NAME,
